Fix clippy warnings, structure, and add build_query to iothub
