+++ src/configuration/modulescontent.rs
@@ -2,45 +2,13 @@ use serde::ser::{Serialize, SerializeStruct, Serializer};
 use serde_json::json;
 use std::collections::HashMap;
 
+use crate::error::{BuilderError, BuilderErrorType};
+
 /// The schema version of the modulescontent
 const SCHEMA_VERSION: &str = "1.0";
 
 /// The runtime type for the containers
 const RUNTIME_TYPE: &str = "docker";
-
-/// Type of the builder error that occurred when building an object
-#[derive(Debug, Clone)]
-pub enum BuilderErrorType {
-    MissingValue(&'static str),
-}
-
-/// BuilderError struct that contains the type of error that occurred
-/// when using a builder
-#[derive(Debug, Clone)]
-pub struct BuilderError {
-    error_type: BuilderErrorType,
-}
-
-impl BuilderError {
-    /// Create a new BuilderError struct
-    pub fn new(error_type: BuilderErrorType) -> Self {
-        BuilderError { error_type }
-    }
-}
-
-impl std::fmt::Display for BuilderError {
-    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
-        match self.error_type {
-            BuilderErrorType::MissingValue(val) => write!(f, "missing field {}", val),
-        }
-    }
-}
-
-impl std::error::Error for BuilderError {
-    fn source(&self) -> Option<&(dyn std::error::Error + 'static)> {
-        None
-    }
-}
 
 /// The status of a module, either Running or Stopped
 #[derive(Debug, PartialEq)]
@@ -123,7 +91,7 @@ pub struct EdgeModule {
     pub create_options: Option<serde_json::Value>,
 }
 
-struct EdgeModuleBuilder {
+pub struct EdgeModuleBuilder {
     module_id: Option<String>,
     version: Option<String>,
     status: Option<Status>,
