+++ src/lib.rs
@@ -3,21 +3,11 @@ extern crate serde_derive;
 
 pub mod configuration;
 pub mod directmethod;
+pub mod error;
 pub mod iothub;
 pub mod query;
 pub mod twin;
 
 pub use iothub::IoTHubService;
-pub use configuration::modulescontent::{ModulesContent, ModulesContentBuilder};
+pub use configuration::modulescontent::{EdgeModuleBuilder, ModulesContent, ModulesContentBuilder};
 use iothub::API_VERSION;
-
-use query::QueryBuilder;
-use twin::TwinManager;
-
-#[cfg(test)]
-mod tests {
-    #[test]
-    fn it_works() {
-        assert_eq!(2 + 2, 4);
-    }
-}
